requirements.txt
fastapi
uvicorn
pymongo
paho-mqtt
python-jose[cryptography]
bcrypt
scikit-learn
